# docs/openapi.yaml
openapi: 3.0.3
info:
  title: sandboxd /v1 API
  version: "0.3.0"
  description: >
    The public, tenant-scoped API for sandboxd. This is the only surface a
    console or external integration should use — the internal /sandbox and
    /wake endpoints are not part of this contract. Scoped by the API token
    (Bearer); when auth is disabled (the single-user default) the whole
    surface is reachable without a token.
servers:
  - url: http://127.0.0.1:9090
    description: Default loopback bind (docker-compose maps SANDBOXD_API_BIND here)
security:
  - bearerAuth: []
  - {} # auth disabled by default (single-user)
...
paths:
...
  /v1/snapshots:
    get:
      operationId: listSnapshots
      tags: [snapshots]
      summary: List snapshots (tenant-scoped)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  snapshots:
                    type: array
                    items: { $ref: "#/components/schemas/Snapshot" }
    post:
      operationId: createSnapshot
      tags: [snapshots]
      summary: Capture a snapshot of a stopped sandbox
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [source_sandbox_id, name]
              properties:
                source_sandbox_id: { type: string }
                name: { type: string }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Snapshot" }
        "409":
          description: Source sandbox is running; stop it first.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/snapshots/{id}:
    parameters: [ { $ref: "#/components/parameters/Id" } ]
    get:
      operationId: getSnapshot
      tags: [snapshots]
      summary: Get a snapshot
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Snapshot" }
        "404": { $ref: "#/components/responses/Error" }
    delete:
      operationId: deleteSnapshot
      tags: [snapshots]
      summary: Delete a snapshot
      responses:
        "204": { description: Deleted }
components:
  securitySchemes:
    cubeRelayKey:
      type: apiKey
      in: header
      name: x-api-key
    cubeBridgeToken:
      type: apiKey
      in: header
      name: x-baarcha-bridge
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema: { type: string }
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            retryable: { type: boolean }
...
    Snapshot:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        status: { type: string }
        source_sandbox_id: { type: string }
        source_app_id: { type: string, description: "App the snapshot was captured from." }
        base_image: { type: string }
        visibility: { type: string }
        size_bytes: { type: integer }
        created_at: { type: string, format: date-time }